Fix for issue SWDEV 165350.
eigen code contained a __global__ function that had a "pass-by-reference" argument.
This is an error condition, that was previously not getting flagged as one. Newer versions fo HIP will error out on it.

The fix is to simply change the "pass-by-reference" argument to "pass-by-value".

This change has not yet made it into the eigen upstream repo